requirements.txt
annotated-types==0.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
docx2txt==0.8
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
filelock==3.15.4
idna==3.7
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
packaging==24.1
phonenumbers==8.13.40
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-docx==1.1.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
shellingham==1.5.4
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
tldextract==5.1.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
